# src/app/Http/Requests/ArtistRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema(
 *     schema="ArtistRequest",
 *     type="object",
 *     title="ArtistRequest",
 *     required={"name"},
 *     properties={
 *         @OA\Property(property="name", type="string", example="Eminem"),
 *     }
 * )
 */
class ArtistRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'name' => 'required|string|max:255',
        ];
    }
}